# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: nGX API
  version: 1.0.0
  description: |
    Email infrastructure purpose-built for AI agents.

    nGX lets you provision inboxes, send and receive email, manage
    drafts with human-in-the-loop review, subscribe to real-time events via
    webhooks or WebSockets, and search across your entire message history.

    ## Authentication
    All API requests (except `/health`) require a Bearer API key in the
    `Authorization` header:
    ```
    Authorization: Bearer am_live_...
    ```
    Keys are scoped — see the `scopes` field on the key endpoints.

    ## Organization And Keys
    Organizations are created during account signup and are discovered at runtime
    via the authenticated API key. `GET /org` returns the current organization
    for that key.

    `POST /keys` creates additional API keys for the current organization.
    It does not create a new organization or perform account registration.
  contact:
    name: nGX Support
    url: https://yourdomain.com
servers:
  - url: https://api.yourdomain.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Tags
# ---------------------------------------------------------------------------
...
paths:
...
  /inboxes/{inboxId}/drafts:
    parameters:
      - $ref: "#/components/parameters/inboxId"
    get:
      tags: [Drafts]
      operationId: listDrafts
      summary: List drafts in an inbox
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/pageToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftList"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Drafts]
      operationId: createDraft
      summary: Create a draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDraftRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /inboxes/{inboxId}/drafts/{draftId}:
    parameters:
      - $ref: "#/components/parameters/inboxId"
      - $ref: "#/components/parameters/draftId"
    get:
      tags: [Drafts]
      operationId: getDraft
      summary: Get a draft
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [Drafts]
      operationId: updateDraft
      summary: Update a draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDraftRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Drafts]
      operationId: deleteDraft
      summary: Delete a draft
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatus"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /inboxes/{inboxId}/drafts/{draftId}/approve:
    parameters:
      - $ref: "#/components/parameters/inboxId"
      - $ref: "#/components/parameters/draftId"
    post:
      tags: [Drafts]
      operationId: approveDraft
      summary: Approve a draft
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /inboxes/{inboxId}/drafts/{draftId}/reject:
    parameters:
      - $ref: "#/components/parameters/inboxId"
      - $ref: "#/components/parameters/draftId"
    post:
      tags: [Drafts]
      operationId: rejectDraft
      summary: Reject a draft
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  # -------------------------------------------------------------------------
  # Labels
  # -------------------------------------------------------------------------
  /labels:
    get:
      tags: [Labels]
      operationId: listLabels
      summary: List labels
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/pageToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelList"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Labels]
      operationId: createLabel
      summary: Create a label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabelRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /labels/{labelId}:
    parameters:
      - $ref: "#/components/parameters/labelId"
    get:
      tags: [Labels]
      operationId: getLabel
      summary: Get a label
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [Labels]
      operationId: updateLabel
      summary: Update a label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLabelRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Labels]
      operationId: deleteLabel
      summary: Delete a label
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatus"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  # -------------------------------------------------------------------------
  # Webhooks
  # -------------------------------------------------------------------------
...
  /keys:
    get:
      tags: [Keys]
      operationId: listAPIKeys
      summary: List API keys
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyList"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Keys]
      operationId: createAPIKey
      summary: Create an API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKeyRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAPIKeyResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "API Key (am_live_...)"
      description: |
        All API requests require a Bearer API key.
        Keys can be scoped to specific capabilities and pods.

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
...
    inboxId:
      name: inboxId
      in: path
      required: true
      description: Inbox ID.
      schema:
        type: string
...
    draftId:
      name: draftId
      in: path
      required: true
      description: Draft ID.
      schema:
        type: string
    labelId:
      name: labelId
      in: path
      required: true
      description: Label ID.
      schema:
        type: string
...
    limit:
      name: limit
      in: query
      description: Maximum number of items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    pageToken:
      name: page_token
      in: query
      description: Opaque page token from a previous response.
      schema:
        type: string

  # -------------------------------------------------------------------------
  # Responses
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Bad Request — invalid parameters or request body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "validation failed"
            details:
              field: "url"
              reason: "must be a valid https URL"
    Unauthorized:
      description: Unauthorized — missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "missing or invalid authorization header"
...
    NotFound:
      description: Not Found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "not found"
    Conflict:
      description: Conflict — resource already exists or state transition is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "email address already registered"
...
    InternalError:
      description: Internal Server Error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "internal server error"

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: "not found"
        details:
          type: object
          description: Optional structured error details (validation errors, etc.).
          additionalProperties: true
      example:
        error: "not found"
        details:
          resource: "thread"
          id: "9b6ba4f5-9f7d-4f56-b0ec-f3a0d5f3a201"
    OperationStatus:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        resource:
          type: string
        id:
          type: string
          format: uuid
      example:
        success: true
        message: "resource deleted"
        resource: "inbox"
        id: "3f2d9d77-89b9-4b8f-9f47-d0b78664b6cb"
...
    AttachmentUpload:
      type: object
      required: [filename, content_type, content]
      properties:
        filename:
          type: string
        content_type:
          type: string
        content:
          type: string
          description: Base64-encoded attachment body.

    # --- Draft ----------------------------------------------------------------
    Draft:
      type: object
      required: [draft_id, inbox_id, created_at, updated_at]
      properties:
        draft_id:
          type: string
        inbox_id:
          type: string
        thread_id:
          type: string
          nullable: true
        labels:
          type: array
          items:
            type: string
        reply_to:
          type: array
          items:
            type: string
        to:
          type: array
          items:
            type: string
        cc:
          type: array
          items:
            type: string
        bcc:
          type: array
          items:
            type: string
        subject:
          type: string
        text:
          type: string
        html:
          type: string
        client_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        draft_id: "draft_123"
        inbox_id: "support@example.com"
        thread_id: "thread_123"
        labels:
          - "urgent"
        reply_to: []
        to:
          - "billing@vendor.com"
        cc: []
        bcc: []
        subject: "Re: Invoice #1234"
        text: "Draft reply body"
        html: "<p>Draft reply body</p>"
        client_id: "client_123"
        created_at: "2026-03-27T18:45:00Z"
        updated_at: "2026-03-27T18:45:00Z"
    DraftList:
      type: object
      required: [count, drafts]
      properties:
        count:
          type: integer
        drafts:
          type: array
          items:
            $ref: "#/components/schemas/Draft"
        limit:
          type: integer
          nullable: true
        next_page_token:
          type: string
          nullable: true
      example:
        count: 1
        drafts:
          - draft_id: "draft_123"
            inbox_id: "support@example.com"
            thread_id: "thread_123"
            labels:
              - "urgent"
            reply_to: []
            to:
              - "billing@vendor.com"
            cc: []
            bcc: []
            subject: "Re: Invoice #1234"
            text: "Draft reply body"
            html: "<p>Draft reply body</p>"
            client_id: "client_123"
            created_at: "2026-03-27T18:45:00Z"
            updated_at: "2026-03-27T18:45:00Z"
        limit: 20
        next_page_token: null
    CreateDraftRequest:
      type: object
      required: [to]
      properties:
        thread_id:
          type: string
        in_reply_to:
          type: string
        subject:
          type: string
        to:
          type: array
          minItems: 1
          items:
            type: string
        cc:
          type: array
          items:
            type: string
        bcc:
          type: array
          items:
            type: string
        reply_to:
          type: array
          items:
            type: string
        add_labels:
          type: array
          items:
            type: string
        text:
          type: string
        html:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentUpload"
    UpdateDraftRequest:
      type: object
      properties:
        subject:
          type: string
        to:
          type: array
          minItems: 1
          items:
            type: string
        cc:
          type: array
          items:
            type: string
        bcc:
          type: array
          items:
            type: string
        reply_to:
          type: array
          items:
            type: string
        add_labels:
          type: array
          items:
            type: string
        remove_labels:
          type: array
          items:
            type: string
        text:
          type: string
        html:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentUpload"

    # --- Label ----------------------------------------------------------------
    Label:
      type: object
      required: [label_id, name, created_at, updated_at]
      properties:
        label_id:
          type: string
        name:
          type: string
          example: "urgent"
        color:
          type: string
          nullable: true
          description: Optional hex color code.
          example: "#ff0000"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        label_id: "lbl_123"
        name: "urgent"
        color: "#ff0000"
        created_at: "2026-03-27T18:00:00Z"
        updated_at: "2026-03-27T18:00:00Z"
    LabelList:
      type: object
      required: [count, labels]
      properties:
        count:
          type: integer
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        limit:
          type: integer
          nullable: true
        next_page_token:
          type: string
          nullable: true
      example:
        count: 1
        labels:
          - label_id: "lbl_123"
            name: "urgent"
            color: "#ff0000"
            created_at: "2026-03-27T18:00:00Z"
            updated_at: "2026-03-27T18:00:00Z"
        limit: 20
        next_page_token: null
    CreateLabelRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "urgent"
        color:
          type: string
          description: Optional hex color code.
          example: "#ff0000"
    UpdateLabelRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        color:
          type: string
          nullable: true

    # --- Webhook --------------------------------------------------------------
...
    APIKey:
      type: object
      required: [api_key_id, name, key_prefix, scopes, created_at]
      properties:
        api_key_id:
          type: string
        name:
          type: string
        key_prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        pod_id:
          type: string
          nullable: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      example:
        api_key_id: "key_123"
        name: "prod-agent-key"
        key_prefix: "am_live_AbCd"
        scopes:
          - "inbox:read"
          - "inbox:write"
        pod_id: "pod_123"
        last_used_at: "2026-03-27T18:25:00Z"
        expires_at: null
        revoked_at: null
        created_at: "2026-03-27T18:00:00Z"
    APIKeyList:
      type: object
      required: [count, api_keys]
      properties:
        count:
          type: integer
        api_keys:
          type: array
          items:
            $ref: "#/components/schemas/APIKey"
        next_page_token:
          type: string
          nullable: true
      example:
        count: 1
        api_keys:
          - api_key_id: "key_123"
            name: "prod-agent-key"
            key_prefix: "am_live_AbCd"
            scopes:
              - "inbox:read"
              - "inbox:write"
            pod_id: "pod_123"
            last_used_at: "2026-03-27T18:25:00Z"
            expires_at: null
            revoked_at: null
            created_at: "2026-03-27T18:00:00Z"
        next_page_token: null
    CreateAPIKeyRequest:
      type: object
      required: [name, scopes]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "prod-agent-key"
        scopes:
          type: array
          minItems: 1
          items:
            type: string
        pod_id:
          type: string
          description: Scope this key to a single pod.
        expires_at:
          type: string
          format: date-time
          description: If set, the key automatically expires at this time.
    CreateAPIKeyResponse:
      type: object
      required: [key, api_key]
      properties:
        key:
          type: string
          description: |
            The full API key value. **Shown only once — store securely.**
          example: "am_live_AbCdEfGhIjKlMnOpQrStUvWxYz0123456789"
        api_key:
          $ref: "#/components/schemas/APIKey"
      example:
        key: "am_live_AbCdEfGhIjKlMnOpQrStUvWxYz0123456789"
        api_key:
          api_key_id: "key_123"
          name: "prod-agent-key"
          key_prefix: "am_live_AbCd"
          scopes:
            - "inbox:read"
            - "inbox:write"
          pod_id: "pod_123"
          last_used_at: null
          expires_at: null
          revoked_at: null
          created_at: "2026-03-27T18:00:00Z"

    # --- Search ---------------------------------------------------------------
...